+++ ChangeLog.md
@@ -25,6 +25,8 @@
 
   * Add course-instance-wide issue reporting flag (Matt West).
 
+  * Add advertisement for HackIllinois 2018 (Matt West).
+
   * Fix broken file upload element (Nathan Walters).
 
   * Fix broken popover and improve assessment label styles (Nathan Walters).